move the test to ExecutorSideSQLConfSuite
